fix screen2

+++ stl_to_json.py
@@ -13,6 +13,6 @@ for i in g:
         temp.append(p.index(j))
     template["surface"].append(str(tuple(temp)))
 template["color"] = ["(58,100,100)" for i in range(len(template["surface"]))]
-template["screen"] = 700*50/max([i[0] for i in p])
+template["screen"] = int(700*50/max([i[0] for i in p]))
 with open("./stl.json", "w", encoding="utf-8") as f:
     json.dump(template, f, indent=2)
